LayoutTests/discovery/video-player-has-video-track-expected.txt

Tests HTTPMediaElement Player.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Video Track added. selected: true
PASS vid.videoTracks.length>0 is true
PASS vid.videoTracks.length>0 is true
PASS vid.videoTracks.item(0).selected is true
PASS successfullyParsed is true

TEST COMPLETE
